Add Indy client VM and fix configurations to get it working with NFS VM
Previously this was a single VM project that exported one NFS volume.
However, we needed something more targeted at supporting an Indy
deployment with the NFS volume used for storage.

This modification creates that configuration with flexible scripts
that can be used to fine-tune performance aspects of the NFS mount,
along with the specific Indy to use.

Finally, I added more documentation on how to use this Vagrant config.