Improve Indy restart functions with NFS mounting at a higher location. Add report checking to multibuild.
Previously we were just mounting NFS for artifact storage. This meant that tracking reports wouldn't
be carried over a vagrant VM restart on the client (Indy) VM. However, changing to mount NFS one
directory higher meant that the UI files would have to be copied in from the new Indy tarball
to the NFS volume in order to preserve the Indy web UI.

In addition, I needed the ability to turn off the NFS mount altogether in an Indy deployment
to test with a local-to-the-VM disk. I added a USE_NFS variable to the start-indy.sh script
for this purpose.

To make the VMs more stable on startup, I added network targets to the systemd scripts,
to make the NFS and Indy startups wait until a valid network connection exists.

Finally, I added a reporter function to multibuild.py, which is an adaptation
of the check-report.py script I wrote earlier. It performs the same basic
steps, and writes the file out to the builds/<tracking-id> directory, to
a JSON file.