Moving vagrant-specific parts into vagrant/ subdir
Vagrant synchronizes the whole directory structure where the
Vagrantfile is in the root. In our case, this can mean quite a
lot of files, as temporary content downloads pile up from
successive test runs. The test directory structures need to be
separated from this synchronization process, especially since
they don't add any value to the running VMs.

Pushing all stuff related to running the VMs into a vagrant/ directory
to establish a new root directory for syncronizing to the VMs.