More closely mimic Project Newcastle build process, add tighter options for report checking
PNC does a lot more than just run the build, and some of these steps could
conceivably interfere with successive build runs. This change updates
the process followed by multibuild to mimic all the steps PNC follows.

Also, the separate report verification command multicheck used to try
to check all trackingIDs on the system, which can get to be a large
set. Update this to allow scanning a builds* directory structure and
only verify the tracking reports for the builds in that test execution.